+++ src/index.ts
@@ -1,5 +1,6 @@
 export { closest } from "./closest";
 export { setAttribute } from './set_attribute';
+export { getAttribute } from './get_attribute';
 export { querySelector } from "./query_selector";
 export { createElement } from "./create_element";
 export { createTextNode } from "./create_text_node";